# filters/diag: speedup latency filter tests