--- a/filters/diag/export_test.go
+++ b/filters/diag/export_test.go
@@ -11,6 +11,8 @@ func SetSleep(f filters.Filter, sleep func(time.Duration)) {
 	switch f := f.(type) {
 	case *histFilter:
 		f.sleep = sleep
+	case *jitter:
+		f.sleep = sleep
 	default:
 		panic(fmt.Sprintf("unsupported filter type: %T", f))
 	}