Use mocked sleep (similar to #2432) to speedup latency filters tests.

Before:
```
$ go test ./filters/diag/ -v -run='TestRequestLatency|TestResponseLatency' | grep PASS
--- PASS: TestRequestLatency (21.06s)
    --- PASS: TestRequestLatency/test_uniform_latency (10.54s)
    --- PASS: TestRequestLatency/test_normal_latency (10.52s)
--- PASS: TestResponseLatency (21.40s)
    --- PASS: TestResponseLatency/test_response_uniform_latency (10.60s)
    --- PASS: TestResponseLatency/test_response_normal_latency (10.79s)
PASS
```

After:
```
$ go test ./filters/diag/ -v -run='TestRequestLatency|TestResponseLatency' | grep PASS
--- PASS: TestRequestLatency (0.00s)
    --- PASS: TestRequestLatency/test_uniform_latency (0.00s)
    --- PASS: TestRequestLatency/test_normal_latency (0.00s)
--- PASS: TestResponseLatency (0.00s)
    --- PASS: TestResponseLatency/test_response_uniform_latency (0.00s)
    --- PASS: TestResponseLatency/test_response_normal_latency (0.00s)
PASS
```